vercel-bot/requirements.txt
flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
gspread==6.1.2
google-auth==2.29.0
python-telegram-bot
